requirements.txt
pillow
pywebio
flask
qrcode
python-barcode

